+++ samples/cmds.py
@@ -26,6 +26,11 @@ from pysui.sui.sui_builders import (
     SenderEventQuery,
     TransactionEventQuery,
     TimeRangeEventQuery,
+    GetTxsFromAddress,
+    GetTxsToAddress,
+    GetTxsInputObject,
+    GetTxsMutateObject,
+    GetTxsMoveFunction,
 )
 from pysui.sui.sui_constants import SUI_COIN_DENOMINATOR
 from pysui.sui.sui_types import SuiBoolean, SuiMap, SuiString, EventID
@@ -313,7 +318,7 @@ def events_all(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args = vars(args)
     result = wallet.get_events(**_convert_event_query(var_args, SuiString("All")))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -326,7 +331,7 @@ def events_module(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("module")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -338,7 +343,7 @@ def events_struct(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("struct_name")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -350,7 +355,7 @@ def events_object(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("object")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -362,7 +367,7 @@ def events_recipient(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("recipient")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -374,7 +379,7 @@ def events_sender(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("sender")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -387,7 +392,7 @@ def events_time(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("end_time")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -399,7 +404,7 @@ def events_tx(wallet: SuiWallet, args: argparse.Namespace) -> None:
     var_args.pop("digest")
     result = wallet.get_events(**_convert_event_query(var_args, query))
     if result.is_ok():
-        print(json.dumps(result.result_data, indent=2))
+        print(result.result_data.to_json(indent=2))
     else:
         print(f"Error: {result.result_string}")
 
@@ -422,6 +427,89 @@ def txn_txn(wallet: SuiWallet, args: argparse.Namespace) -> None:
         print(f"Error: {result.result_string}")
 
 
+def _convert_txns_query(var_args: argparse.Namespace, query: Union[SuiString, SuiMap]) -> dict:
+    """Convert arguments to GetTxns."""
+    var_args.pop("version")
+    var_args["query"] = query
+    var_args["cursor"] = SuiString(var_args["digest"])
+    var_args.pop("digest")
+    var_args["descending_order"] = SuiBoolean(var_args["descending_order"])
+    return var_args
+
+
+def txns_all(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    result = wallet.get_txns(**_convert_txns_query(var_args, SuiString("All")))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+
+
+def txns_movefunc(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    query = GetTxsMoveFunction(args.package.value, args.function, args.module)
+    var_args.pop("package")
+    var_args.pop("function")
+    var_args.pop("module")
+    result = wallet.get_txns(**_convert_txns_query(var_args, query))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+    #
+
+
+def txns_input(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    query = GetTxsInputObject(var_args["input"].value)
+    var_args.pop("input")
+    result = wallet.get_txns(**_convert_txns_query(var_args, query))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+
+
+def txns_mutate(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    query = GetTxsMutateObject(var_args["mutated"].value)
+    var_args.pop("mutated")
+    result = wallet.get_txns(**_convert_txns_query(var_args, query))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+
+
+def txns_from(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    query = GetTxsFromAddress(args.froms.value.value)
+    var_args.pop("froms")
+    result = wallet.get_txns(**_convert_txns_query(var_args, query))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+
+
+def txns_to(wallet: SuiWallet, args: argparse.Namespace) -> None:
+    """Transaction information request handler."""
+    var_args = vars(args)
+    query = GetTxsToAddress(args.to.value.value)
+    var_args.pop("to")
+    result = wallet.get_txns(**_convert_txns_query(var_args, query))
+    if result.is_ok():
+        print(result.result_data.to_json(indent=2))
+    else:
+        print(f"Error: {result.result_string}")
+
+
 SUI_CMD_DISPATCH = {
     "event-all": events_all,
     "event-module": events_module,
@@ -433,6 +521,12 @@ SUI_CMD_DISPATCH = {
     "event-tx": events_tx,
     "txn-count": txn_count,
     "txn-txn": txn_txn,
+    "txnsq-all": txns_all,
+    "txnsq-movefunc": txns_movefunc,
+    "txnsq-input": txns_input,
+    "txnsq-mutate": txns_mutate,
+    "txnsq-from": txns_from,
+    "txnsq-to": txns_to,
     "active-address": sui_active_address,
     "addresses": sui_addresses,
     "gas": sui_gas,